event fix

+++ services/model-service.js
@@ -1,5 +1,7 @@
 var models = require('../models');
 
+let publishToOther = null;
+
 const typeMapping = (type) => {
     let output = ''
     switch (type) {
@@ -26,7 +28,7 @@ const typeMapping = (type) => {
 
         //STRING
         case '5':
-            output = "VARCHAR(20)"
+            output = "VARCHAR(10)"
             break;
 
         //CHAR
@@ -129,10 +131,10 @@ const ddlGenerator = (model, devMac, events) => {
                             )
                             VALUES(
                                 ${events[i].eventId},
-                                ${events[i].dataKey},
+                                '${events[i].dataKey}',
                                 ${events[i].outputType},
                                 ${events[i].ruleType},
-                                ${events[i].ruleValue},
+                                ${events[i].ruleValue ? events[i].ruleValue : 'NULL'},
                                 ${events[i].priority}
                             )
                             `
@@ -150,9 +152,9 @@ const ddlGenerator = (model, devMac, events) => {
                                 )
                                 VALUES(
                                     ${events[i].eventId},
-                                    ${events[i].host},
-                                    ${events[i].port},
-                                    ${events[i].path}
+                                    '${events[i].host}',
+                                    '${events[i].port}',
+                                    '${events[i].path}'
                                 )
                                 `
                     );
@@ -168,7 +170,7 @@ const ddlGenerator = (model, devMac, events) => {
                                 VALUES(
                                     ${events[i].eventId},
                                     ${events[i].evCode},
-                                    ${events[i].authKey}
+                                    '${events[i].authKey}'
                                 )
                                 `
                     );
@@ -188,10 +190,41 @@ const dropTable = (tableName) => {
 }
 
 //이벤트 실행 함수
-const actEvent = (type) => {
-    switch(type){
+const actEvent = (type, replaceMac, mac, eventId) => {
+    const controlTable = 'control' + '_' + replaceMac
+    switch (type) {
         case '2':
             console.log('ctrl')
+            models.sequelize.query(
+                `SELECT * 
+                FROM ${controlTable}
+                WHERE event_id=${eventId}`
+            ).then((rows) => {
+                const authKey = rows[0][0].auth_key
+                const evCode = rows[0][0].ev_code;
+                const cmdList = [
+                    {
+                        e: evCode,
+                        t: 1,
+                        d: []
+                    }
+                ]
+
+                //디바이스 등록 여부 조회
+                models.dev.findAll({
+                    attributes: ['dev_mac', 'dev_type'],
+                    where: {
+                        dev_type: authKey
+                    }
+                }).then(devInfo => {
+                    //동일한 타입의 모든 디바이스에게 명령 전송
+                    for (let i = 0; i < devInfo.length; i++) {
+                        publishToOther(devInfo[i].dev_mac, {
+                            c: cmdList
+                        })
+                    }
+                })
+            })
             break;
         case '3':
             console.log('third')
@@ -201,7 +234,7 @@ const actEvent = (type) => {
     }
 }
 
-const checkEventRule = (macStr, key, val) => {
+const checkEventRule = (mac, macStr, key, val) => {
     models.sequelize.query(
         `SELECT * 
         FROM ${'event' + '_' + macStr}
@@ -211,16 +244,16 @@ const checkEventRule = (macStr, key, val) => {
         let isMatch = false;
         for(let i=0; i<rows[0].length; i++){
             switch(rows[0][i].rule_type){
-                case '1':
+                case '2':
                     isMatch = eval(`${val}===${rows[0][i].rule_value}`)
                     break;
-                case '2':
+                case '3':
                     isMatch = eval(`${val}!==${rows[0][i].rule_value}`)
                     break;
-                case '3':
+                case '4':
                     isMatch = eval(`${val}>${rows[0][i].rule_value}`)
                     break;
-                case '4':
+                case '5':
                     isMatch = eval(`${val}<${rows[0][i].rule_value}`)
                     break;
                 //case '0'
@@ -229,7 +262,7 @@ const checkEventRule = (macStr, key, val) => {
                     break;
             }
             if(isMatch){
-                actEvent(rows[0][i].output_type)
+                actEvent(rows[0][i].output_type, macStr, mac, rows[0][i].event_id)
                 return;
             }
         }
@@ -250,10 +283,14 @@ module.exports = {
     deleteDataModelTable: (devMac) => {
         return new Promise((resolve, reject) => {
             const replaceMac = devMac.replace(/:/g, '')
+            let flag = false;
             models.sequelize.query(`SHOW TABLES`).then(async (rows) => {
                 for (let i = 0; i < rows[0].length; i++) {
                     if (rows[0][i].Tables_in_hub_system.indexOf(replaceMac) !== -1) {
-                        await dropTable(rows[0][i].Tables_in_hub_system)
+                        if(rows[0][i].Tables_in_hub_system.indexOf('event') === -1)
+                            await dropTable(rows[0][i].Tables_in_hub_system)
+                        else
+                            flag=true;
                     }
                 }
                 models.sequelize.query('COMMIT')
@@ -280,7 +317,9 @@ module.exports = {
         models.sequelize.query(
             `UPDATE ${record.key + '_' + replaceMac}
             SET data = ${record.val}`
-        );
+        ).then(() => {
+            checkEventRule(record.mac, replaceMac, record.key, record.val)
+        })
     },
 
     getDataModel: (key, mac) => {
@@ -311,7 +350,12 @@ module.exports = {
                 NOW(),
                 ${record.val}
             )`
-        );
-        checkEventRule(replaceMac, record.key, record.val)
+        ).then(() => {
+            checkEventRule(record.mac, replaceMac, record.key, record.val)
+        })
     },
+
+    init: (func) => {
+        publishToOther =func;
+    }
 }
